r3 p3 reusable component, accept input
Let’s take stock of what we’ve built.

We created a reusable component
We learned how to make a component accept input
We learned to declare the application directives we need in an Angular module. We list the directives in the NgModule decorator's declarations array.
We learned to bind a parent component to a child component.